Got rid of static methods
